Fix tests and build issues
