+++ packages/mobile/src/verify/VerificationInputScreen.test.tsx
@@ -7,7 +7,17 @@ import VerificationInputScreen from 'src/verify/VerificationInputScreen'
 import { createMockStore, getMockStackScreenProps } from 'test/utils'
 
 describe('VerificationInputScreen', () => {
-  const store = createMockStore({})
+  const store = createMockStore({
+    verify: {
+      attestationCodes: [],
+      status: {
+        isVerified: false,
+        numAttestationsRemaining: 3,
+        total: 3,
+        completed: 0,
+      },
+    },
+  })
 
   it('renders correctly', () => {
     const { toJSON } = render(
