+++ packages/mobile/src/verify/saga.test.ts
@@ -1,26 +1,23 @@
+import { ActionableAttestation } from '@celo/contractkit/lib/wrappers/Attestations'
 import { verifyWallet } from '@celo/komencikit/src/verifyWallet'
 import { getPhoneHash } from '@celo/utils/lib/phoneNumbers'
 import * as reduxSagaTestPlan from 'redux-saga-test-plan'
 import { throwError } from 'redux-saga-test-plan/providers'
 import { call, delay, select } from 'redux-saga/effects'
+import { VerificationEvents } from 'src/analytics/Events'
+import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { ErrorMessages } from 'src/app/ErrorMessages'
 import networkConfig from 'src/geth/networkConfig'
 import { celoTokenBalanceSelector } from 'src/goldToken/selectors'
-import {
-  setVerificationStatus as setOldVerificationStatus,
-  updateE164PhoneNumberSalts,
-} from 'src/identity/actions'
+import { updateE164PhoneNumberSalts } from 'src/identity/actions'
 import { KomenciErrorQuotaExceeded } from 'src/identity/feelessVerificationErrors'
 import { fetchPhoneHashPrivate } from 'src/identity/privateHashing'
 import { e164NumberToSaltSelector } from 'src/identity/reducer'
-import { VerificationStatus } from 'src/identity/types'
-import { BALANCE_CHECK_TIMEOUT, getActionableAttestations } from 'src/identity/verification'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { waitFor } from 'src/redux/sagas-helpers'
 import { stableTokenBalanceSelector } from 'src/stableToken/reducer'
 import {
-  doVerificationFlow,
   e164NumberSelector,
   ensureRealHumanUser,
   fail,
@@ -30,18 +27,20 @@ import {
   isBalanceSufficientForSigRetrievalSelector,
   KomenciAvailable,
   komenciContextSelector,
-  overrideWithoutVerificationSelector,
   phoneHashSelector,
+  reportRevealStatus,
+  requestAttestations,
   setActionableAttestation,
   setKomenciAvailable,
   setKomenciContext,
-  setOverrideWithoutVerification,
   setPhoneHash,
   setVerificationStatus,
   shouldUseKomenciSelector,
   start,
 } from 'src/verify/reducer'
 import {
+  AttestationCode,
+  BALANCE_CHECK_TIMEOUT,
   checkIfKomenciAvailableSaga,
   failSaga,
   fetchKomenciReadiness,
@@ -49,13 +48,18 @@ import {
   fetchOnChainDataSaga,
   fetchOrDeployMtwSaga,
   fetchPhoneNumberDetailsSaga,
+  getActionableAttestations,
   getKomenciKit,
+  getPhoneHashDetails,
+  reportActionableAttestationsStatuses,
+  reportRevealStatusSaga,
   resetSaga,
   startSaga,
 } from 'src/verify/saga'
 import { getContractKit, getContractKitAsync } from 'src/web3/contracts'
 import { registerWalletAndDekViaKomenci } from 'src/web3/dataEncryptionKey'
 import { getAccount, getConnectedUnlockedAccount, unlockAccount, UnlockResult } from 'src/web3/saga'
+import { mockE164NumberHash, mockE164NumberPepper, mockPublicDEK } from 'test/values'
 
 export const mockAccount = '0x0000000000000000000000000000000000007E57'
 export const mockAccount1 = '0x0000000000000000000000000000000000007E58'
@@ -69,8 +73,14 @@ export const mockKomenciContext = {
   captchaToken: '',
 }
 export const mockE164Number = '+14155550000'
-export const mockPepper = 'pepper'
-export const mockPhoneHash = getPhoneHash(mockE164Number, mockPepper)
+export const mockPhoneHash = getPhoneHash(mockE164Number, mockE164NumberPepper)
+export const mockPhoneHashDetails = {
+  e164Number: mockE164Number,
+  phoneHash: mockPhoneHash,
+  pepper: mockE164NumberPepper,
+}
+
+const MockedAnalytics = ValoraAnalytics as any
 
 const mockActionableAttestation = {
   issuer: mockAccount2,
@@ -80,14 +90,63 @@ const mockActionableAttestation = {
   version: '1.0.0',
 }
 
+const attestationCode0: AttestationCode = {
+  code:
+    'ab8049b95ac02e989aae8b61fddc10fe9b3ac3c6aebcd3e68be495570b2d3da15aabc691ab88de69648f988fab653ac943f67404e532cfd1013627f56365f36501',
+  issuer: '848920b14154b6508b8d98e7ee8159aa84b579a4',
+}
+
+const attestationCode1: AttestationCode = {
+  code:
+    '2033a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7998e1e8b183cf6b8cb4d4a8d923fecfddd191e61e074adc5e00',
+  issuer: 'fdb8da92c3597e81c2737e8be793bee9f1172045',
+}
+
+const attestationCode2: AttestationCode = {
+  code:
+    '1930a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7993f75ab183cf6b8cb4d4a8d923fecfddd191e61e074adc5a10',
+  issuer: 'ecb8da92c3597e81c2737e8be793bee9f1173156',
+}
+
+const mockActionableAttestations: ActionableAttestation[] = [
+  {
+    issuer: attestationCode0.issuer,
+    blockNumber: 100,
+    attestationServiceURL: 'https://fake.celo.org/0',
+    name: '',
+    version: '1.1.0',
+  },
+  {
+    issuer: attestationCode1.issuer,
+    blockNumber: 110,
+    attestationServiceURL: 'https://fake.celo.org/1',
+    name: '',
+    version: '1.1.0',
+  },
+  {
+    issuer: attestationCode2.issuer,
+    blockNumber: 120,
+    attestationServiceURL: 'https://fake.celo.org/2',
+    name: '',
+    version: '1.1.0',
+  },
+]
+
 const mockKomenciKit = {
   getDistributedBlindedPepper: jest.fn(),
   deployWallet: jest.fn(),
 }
 
+const mockAccountsWrapper = {
+  getWalletAddress: jest.fn(() => Promise.resolve(mockAccount)),
+  getDataEncryptionKey: jest.fn(() => Promise.resolve(mockPublicDEK)),
+}
+
 const mockAttestationsWrapper = {
   lookupAccountsForIdentifier: jest.fn(),
   getVerifiedStatus: jest.fn(),
+  getRevealStatus: jest.fn(),
+  getActionableAttestations: jest.fn(),
 }
 
 describe(checkIfKomenciAvailableSaga, () => {
@@ -110,7 +169,6 @@ describe(checkIfKomenciAvailableSaga, () => {
 
 describe(startSaga, () => {
   it('starts with Komenci active session', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     const activeSessionMockKomenciContext = {
       ...mockKomenciContext,
@@ -118,14 +176,9 @@ describe(startSaga, () => {
     }
     const komenciKit = getKomenciKit(contractKit, mockAccount, activeSessionMockKomenciContext)
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: withoutRevealing })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(getContractKit), contractKit],
         [call(getConnectedUnlockedAccount), mockAccount],
         [call(unlockAccount, mockAccount, true), UnlockResult.SUCCESS],
@@ -143,18 +196,12 @@ describe(startSaga, () => {
   })
 
   it('starts with Komenci inactive session', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: withoutRevealing })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(getContractKit), contractKit],
         [call(getConnectedUnlockedAccount), mockAccount],
         [call(unlockAccount, mockAccount, true), UnlockResult.SUCCESS],
@@ -169,18 +216,12 @@ describe(startSaga, () => {
   })
 
   it('disables Komenci if unrecoverable error has occured', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: { withoutRevealing } })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(getContractKit), contractKit],
         [call(getConnectedUnlockedAccount), mockAccount],
         [call(unlockAccount, mockAccount, true), UnlockResult.SUCCESS],
@@ -194,22 +235,16 @@ describe(startSaga, () => {
         ],
       ])
       .put(setKomenciAvailable(KomenciAvailable.No))
-      .put(start({ e164Number: mockE164Number, withoutRevealing }))
+      .put(start({ e164Number: mockE164Number }))
       .run()
   })
 
   it('fails if komenci is disabled and balance fetch timeouts', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: withoutRevealing })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(getContractKit), contractKit],
         [call(getConnectedUnlockedAccount), mockAccount],
         [call(unlockAccount, mockAccount, true), UnlockResult.SUCCESS],
@@ -222,17 +257,11 @@ describe(startSaga, () => {
   })
 
   it('fails if komenci is disabled and balance is insufficient', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: withoutRevealing })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(waitFor, stableTokenBalanceSelector), 1],
         [call(waitFor, celoTokenBalanceSelector), 1],
         [delay(BALANCE_CHECK_TIMEOUT), true],
@@ -248,17 +277,11 @@ describe(startSaga, () => {
   })
 
   it('starts with Komenci disabled', async () => {
-    const withoutRevealing = true
     const contractKit = await getContractKitAsync()
     await reduxSagaTestPlan
-      .expectSaga(startSaga, { payload: withoutRevealing })
+      .expectSaga(startSaga)
       .provide([
-        [
-          call(navigate, Screens.VerificationLoadingScreen, {
-            withoutRevealing,
-          }),
-          null,
-        ],
+        [call(navigate, Screens.VerificationLoadingScreen), null],
         [call(waitFor, stableTokenBalanceSelector), 1],
         [call(waitFor, celoTokenBalanceSelector), 1],
         [delay(BALANCE_CHECK_TIMEOUT), true],
@@ -286,7 +309,7 @@ describe(fetchPhoneNumberDetailsSaga, () => {
         [select(e164NumberSelector), mockE164Number],
         [select(shouldUseKomenciSelector), true],
         [select(phoneHashSelector), mockPhoneHash],
-        [select(e164NumberToSaltSelector), { [mockE164Number]: mockPepper }],
+        [select(e164NumberToSaltSelector), { [mockE164Number]: mockE164NumberPepper }],
       ])
       .put(fetchMtw())
       .run()
@@ -303,7 +326,7 @@ describe(fetchPhoneNumberDetailsSaga, () => {
         [select(e164NumberSelector), mockE164Number],
         [select(shouldUseKomenciSelector), true],
         [select(phoneHashSelector), null],
-        [select(e164NumberToSaltSelector), { [mockE164Number]: mockPepper }],
+        [select(e164NumberToSaltSelector), { [mockE164Number]: mockE164NumberPepper }],
       ])
       .put(fetchMtw())
       .put(setPhoneHash(mockPhoneHash))
@@ -315,7 +338,7 @@ describe(fetchPhoneNumberDetailsSaga, () => {
     const komenciKit = mockKomenciKit // getKomenciKit(contractKit, mockAccount, mockKomenciContext)
     ;(komenciKit.getDistributedBlindedPepper as jest.Mock).mockReturnValueOnce({
       ok: true,
-      result: { pepper: mockPepper },
+      result: { pepper: mockE164NumberPepper },
     })
     await reduxSagaTestPlan
       .expectSaga(fetchPhoneNumberDetailsSaga)
@@ -330,7 +353,7 @@ describe(fetchPhoneNumberDetailsSaga, () => {
         [call(getKomenciKit, contractKit, mockAccount, mockKomenciContext), komenciKit],
         [select(komenciContextSelector), mockKomenciContext],
       ])
-      .put(updateE164PhoneNumberSalts({ [mockE164Number]: mockPepper }))
+      .put(updateE164PhoneNumberSalts({ [mockE164Number]: mockE164NumberPepper }))
       .put(setPhoneHash(mockPhoneHash))
       .put(fetchMtw())
       .run()
@@ -350,11 +373,11 @@ describe(fetchPhoneNumberDetailsSaga, () => {
         [select(e164NumberToSaltSelector), {}],
         [
           call(fetchPhoneHashPrivate, mockE164Number),
-          { pepper: mockPepper, phoneHash: mockPhoneHash },
+          { pepper: mockE164NumberPepper, phoneHash: mockPhoneHash },
         ],
         [select(komenciContextSelector), mockKomenciContext],
       ])
-      .put(updateE164PhoneNumberSalts({ [mockE164Number]: mockPepper }))
+      .put(updateE164PhoneNumberSalts({ [mockE164Number]: mockE164NumberPepper }))
       .put(setPhoneHash(mockPhoneHash))
       .put(fetchOnChainData())
       .run()
@@ -566,7 +589,6 @@ describe(fetchOrDeployMtwSaga, () => {
           isVerified: true,
         })
       )
-      .put(doVerificationFlow(true))
       .run()
   })
 })
@@ -591,7 +613,6 @@ describe(fetchOnChainDataSaga, () => {
         [select(e164NumberSelector), mockE164Number],
         [call(getContractKit), contractKit],
         [select(komenciContextSelector), mockKomenciContextWithUnverifiedMtwAddress],
-        [select(overrideWithoutVerificationSelector), true],
         [select(shouldUseKomenciSelector), true],
         [select(phoneHashSelector), mockPhoneHash],
         [
@@ -604,13 +625,12 @@ describe(fetchOnChainDataSaga, () => {
         },
       ])
       .put(setActionableAttestation([mockActionableAttestation]))
-      .put(setOverrideWithoutVerification(undefined))
       .put(
         setVerificationStatus({
           isVerified: false,
         })
       )
-      .put(doVerificationFlow(true))
+      .put(requestAttestations())
       .run()
   })
 
@@ -633,7 +653,6 @@ describe(fetchOnChainDataSaga, () => {
         [select(e164NumberSelector), mockE164Number],
         [call(getContractKit), contractKit],
         [select(komenciContextSelector), mockKomenciContextWithUnverifiedMtwAddress],
-        [select(overrideWithoutVerificationSelector), true],
         [select(shouldUseKomenciSelector), false],
         [select(phoneHashSelector), mockPhoneHash],
         [call(getAccount), mockAccount],
@@ -647,13 +666,12 @@ describe(fetchOnChainDataSaga, () => {
         },
       ])
       .put(setActionableAttestation([mockActionableAttestation]))
-      .put(setOverrideWithoutVerification(undefined))
       .put(
         setVerificationStatus({
           isVerified: false,
         })
       )
-      .put(doVerificationFlow(true))
+      .put(requestAttestations())
       .run()
   })
   it('emits fail event if something goes wrong', async () => {
@@ -679,7 +697,103 @@ describe(failSaga, () => {
   it('set verification status to failed', async () => {
     await reduxSagaTestPlan
       .expectSaga(failSaga, 'test')
-      .put(setOldVerificationStatus(VerificationStatus.Failed))
+      .provide([[select(shouldUseKomenciSelector), true]])
+      .run()
+  })
+})
+
+describe(reportRevealStatusSaga, () => {
+  beforeEach(() => {
+    MockedAnalytics.track.mockReset()
+  })
+  it('report actionable attestation to analytics', async () => {
+    const contractKit = await getContractKitAsync()
+    ;(mockAttestationsWrapper.getRevealStatus as jest.Mock).mockReturnValue({
+      ok: true,
+      json: () => body,
+    })
+
+    const mockIssuer = mockActionableAttestation.issuer
+    const body = { issuer: mockIssuer, custom: 'payload' }
+    await reduxSagaTestPlan
+      .expectSaga(reportRevealStatusSaga, {
+        payload: {
+          attestationServiceUrl: 'url',
+          e164Number: mockE164Number,
+          account: mockAccount,
+          issuer: mockIssuer,
+          pepper: mockE164NumberPepper,
+        },
+      })
+      .provide([
+        [
+          call([contractKit.contracts, contractKit.contracts.getAttestations]),
+          mockAttestationsWrapper,
+        ],
+      ])
+      .run()
+    expect(MockedAnalytics.track.mock.calls.length).toBe(1)
+    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(
+      VerificationEvents.verification_reveal_attestation_status
+    )
+    expect(MockedAnalytics.track.mock.calls[0][1]).toStrictEqual(body)
+  })
+})
+
+describe(reportActionableAttestationsStatuses, () => {
+  it('report actionable attestations', async () => {
+    const contractKit = await getContractKitAsync()
+    ;(mockAttestationsWrapper.getActionableAttestations as jest.Mock).mockReturnValue(
+      mockActionableAttestations
+    )
+
+    await reduxSagaTestPlan
+      .expectSaga(reportActionableAttestationsStatuses)
+      .provide([
+        [call(getConnectedUnlockedAccount), mockAccount],
+        [call(getPhoneHashDetails), mockPhoneHashDetails],
+        [select(e164NumberSelector), mockE164Number],
+        [
+          call([contractKit.contracts, contractKit.contracts.getAttestations]),
+          mockAttestationsWrapper,
+        ],
+        [call([contractKit.contracts, contractKit.contracts.getAccounts]), mockAccountsWrapper],
+        [
+          call(fetchPhoneHashPrivate, mockE164Number),
+          {
+            phoneHash: mockE164NumberHash,
+            e164Number: mockE164Number,
+            pepper: mockE164NumberPepper,
+          },
+        ],
+      ])
+      .put(
+        reportRevealStatus({
+          attestationServiceUrl: mockActionableAttestations[0].attestationServiceURL,
+          account: mockAccount,
+          issuer: mockActionableAttestations[0].issuer,
+          e164Number: mockE164Number,
+          pepper: mockE164NumberPepper,
+        })
+      )
+      .put(
+        reportRevealStatus({
+          attestationServiceUrl: mockActionableAttestations[1].attestationServiceURL,
+          account: mockAccount,
+          issuer: mockActionableAttestations[1].issuer,
+          e164Number: mockE164Number,
+          pepper: mockE164NumberPepper,
+        })
+      )
+      .put(
+        reportRevealStatus({
+          attestationServiceUrl: mockActionableAttestations[2].attestationServiceURL,
+          account: mockAccount,
+          issuer: mockActionableAttestations[2].issuer,
+          e164Number: mockE164Number,
+          pepper: mockE164NumberPepper,
+        })
+      )
       .run()
   })
 })
